requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pycryptodome==3.20.0
dmPython>=2.4
redis==5.0.0
loguru==0.7.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
